==> notebooks/01_fetch_data.ipynb <==
from pathlib import Path
import pandas as pd
import requests

def fetch_raw_data(year: int, month: int) -> str:
    url = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02}.parquet"
    response = requests.get(url)
    if response.status_code == 200:
        # Use pathlib to construct the path
        path = Path("..") / "data" / "raw" / f"rides_{year}_{month:02}.parquet"
        # Ensure the parent directory exists
        path.parent.mkdir(parents=True, exist_ok=True)
        # Write the content to the file
        path.write_bytes(response.content)
        print(f"Successfully fetched: {str(path)}")
        return str(path)
    else:
        raise Exception(f"{url} is not available")

fetch_raw_data(2023, 1)
