==> notebooks/05_transform_raw_data_into_features_and_targets.ipynb <==
%load_ext autoreload
%autoreload 2

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

# Now you can import from src
from src.data_utils import load_and_process_taxi_data

rides = load_and_process_taxi_data(year=2023)

rides

from src.data_utils import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)
ts_data.head()

ts_data.shape

from src.data_utils import transform_ts_data_info_features_and_target_loop

features, targets = transform_ts_data_info_features_and_target_loop(ts_data, window_size=24*28, step_size=24)

features

import numpy as np
window_size = 4
step_size = 1
num_windows = 4

step_size * np.arange(num_windows)[:, None]  # Output: array([[0], [1], [2], [3]])

indices = np.arange(window_size)[None, :] + step_size * np.arange(num_windows)[:, None]

window_size = 4
step_size = 1
num_windows = 4

from src.data_utils import transform_ts_data_info_features_and_target_loop

features, targets = transform_ts_data_info_features_and_target_loop(ts_data, window_size=24*28*1, step_size=24)

features

import numpy as np
np.arange(num_windows)

from src.data_utils import transform_ts_data_info_features_and_target_loop

features, targets = transform_ts_data_info_features_and_target_loop(ts_data, window_size=24*28*1, step_size=24)


tabular_data = features
tabular_data["target"] = targets

from src.config import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet", engine="pyarrow")

features, targets = transform_ts_data_info_features_and_target_loop(ts_data[ts_data["pickup_location_id"]==43], window_size=12, step_size=1)


ts_data[ts_data["pickup_location_id"]==43].head(36)

from src.data_utils import transform_ts_data_info_features_and_target

features, targets = transform_ts_data_info_features_and_target(ts_data[ts_data["pickup_location_id"]==43], window_size=4, step_size=1)

features

from src.data_utils import transform_ts_data_info_features_and_target
features, targets = transform_ts_data_info_features_and_target(ts_data[ts_data["pickup_location_id"]==43], window_size=4, step_size=1)
features
